Conflicts:
	lib/FusionInventory/Agent.pm
	lib/FusionInventory/Agent/Config.pm
	lib/FusionInventory/Agent/Task.pm
	lib/FusionInventory/Agent/Task/Inventory.pm
	lib/FusionInventory/Agent/Task/Inventory/OS/AIX/Drives.pm
	lib/FusionInventory/Agent/Tools.pm
	lib/FusionInventory/Agent/Tools/Linux.pm

+++ lib/FusionInventory/Logger/File.pm
@@ -32,7 +32,7 @@ sub addMsg {
 
     if ($self->{logfile_maxsize}) {
         my $stat = stat($self->{logfile});
-        if ($stat->size() > $self->{logfile_maxsize}) {
+        if ($stat && $stat->size() > $self->{logfile_maxsize}) {
             unlink $self->{logfile}
                 or warn "Can't unlink $self->{logfile}: $ERRNO";
         }
